third_party/WebKit/LayoutTests/external/wpt/service-workers/service-worker/fetch-event.https-expected.txt

This is a testharness.js-based test.
PASS Service Worker responds to fetch event with string 
PASS Service Worker responds to fetch event with blob body 
PASS Service Worker responds to fetch event with the referrer URL 
FAIL Service Worker responds to fetch event with the referrer policy assert_unreached: unexpected rejection: assert_equals: Service Worker should respond to fetch with the referrer URL when a member of RequestInit is present expected "Referrer: https://web-platform.test:8444/service-workers/service-worker/resources/simple.html?referrerPolicy\nReferrerPolicy: no-referrer-when-downgrade" but got "Referrer: https://web-platform.test:8444/service-workers/service-worker/resources/undefined\nReferrerPolicy: no-referrer-when-downgrade" Reached unreachable code
PASS Service Worker responds to fetch event with an existing client id 
PASS Service Worker does not respond to fetch event 
PASS Service Worker responds to fetch event with null response body 
PASS Service Worker fetches other file in fetch event 
PASS Service Worker responds to fetch event with POST form 
PASS Multiple calls of respondWith must throw InvalidStateErrors 
PASS Service Worker event.respondWith must set the used flag 
FAIL Service Worker must not expose FetchEvent URL fragments. assert_unreached: unexpected rejection: assert_equals: Service worker should expose URL fragments in request. expected "Fragment Found :#/some/fragment" but got "Fragment Not Found" Reached unreachable code
FAIL Service Worker responds to fetch event with the correct cache types assert_unreached: unexpected rejection: assert_equals: expected "default" but got "" Reached unreachable code
FAIL Service Worker should intercept EventSource assert_unreached: unexpected rejection: assert_equals: EventSource should bypass the http cache. expected (string) "no-store" but got (undefined) undefined Reached unreachable code
PASS Service Worker responds to fetch event with the correct integrity_metadata 
Harness: the test ran to completion.
